Changed EyeType from byte to short (NA187)

+++ src/ChannelServer/Database/ChannelDb.cs
@@ -180,7 +180,7 @@ namespace Aura.Channel.Database
 					character.Race = reader.GetInt32("race");
 					character.DeletionTime = reader.GetDateTimeSafe("deletionTime");
 					character.SkinColor = reader.GetByte("skinColor");
-					character.EyeType = reader.GetByte("eyeType");
+					character.EyeType = reader.GetInt16("eyeType");
 					character.EyeColor = reader.GetByte("eyeColor");
 					character.MouthType = reader.GetByte("mouthType");
 					character.Height = reader.GetFloat("height");
